+++ R/tm_g_butterfly.R
@@ -262,6 +262,12 @@ srv_g_butterfly <- function(id, datasets, reporter, dataname, label, plot_height
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
 
   moduleServer(id, function(input, output, session) {
+    iv <- shinyvalidate::InputValidator$new()
+    iv$add_rule("category_var", shinyvalidate::sv_required())
+    iv$add_rule("right_var", shinyvalidate::sv_required())
+    iv$add_rule("left_var", shinyvalidate::sv_required())
+    iv$enable()
+
     teal.code::init_chunks()
 
     options <- reactiveValues(r = NULL, l = NULL)
@@ -360,8 +366,7 @@ srv_g_butterfly <- function(id, datasets, reporter, dataname, label, plot_height
 
 
     plot_r <- reactive({
-      validate(need(input$category_var, "Please select a category variable."))
-
+      validate(need(iv$is_valid(), "Misspecification error: please observe red flags in the encodings."))
       ADSL <- datasets$get_data("ADSL", filtered = TRUE) # nolint
       ANL <- datasets$get_data(dataname, filtered = TRUE) # nolint
 
@@ -377,17 +382,15 @@ srv_g_butterfly <- function(id, datasets, reporter, dataname, label, plot_height
       sort_by_var <- input$sort_by_var
       filter_var <- input$filter_var
 
+      iv_len <- shinyvalidate::InputValidator$new()
+      iv_len$add_rule("right_val", shinyvalidate::sv_required("Please select at least one"))
+      iv_len$add_rule("left_val", shinyvalidate::sv_required("Please select at least one"))
+      iv_len$enable()
+      validate(need(iv_len$is_valid(), "Misspecification error: please observe red flags in the encodings."))
+
       validate(
         need(nrow(ADSL) > 0, "ADSL Data has no rows"),
         need(nrow(ANL) > 0, "ADAE Data has no rows")
-      )
-      validate(
-        need(right_var, "'Right Dichotomization Variable' not selected"),
-        need(left_var, "'Left Dichotomization Variable' not selected")
-      )
-      validate(
-        need(length(right_val) > 0, "No values of 'Right Dichotomization Variable' are checked"),
-        need(length(left_val) > 0, "No values of 'Left Dichotomization Variable' are checked")
       )
       validate(need(
         any(c(ADSL[[right_var]] %in% right_val, ADSL[[left_var]] %in% left_val)),